Weight pic days by a normal distribution kernel
This makes it more likely to pick pictures from the current month
or surrounding months, but very unlikely to pick pictures from 6 months
away.

We could probably make this more fine-tuned to apply the weighting by week
instead of month, but probably don't want to go finer than that. If
we did it by day, for example, we might totally miss days where we did
not have the screensaver on.