Cache images in the current direction
It seems that loading the files from disk makes changing the picture seem sluggish. In this change, we preload the image in the direction we are going (e.g., if we are stepping forward in time, we preload the next chronological picture). We also always preload the next random image.